% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{convert_coordinates}
\alias{convert_coordinates}
\title{Convert Genomic Coordinates Between Genome Assemblies}
\usage{
convert_coordinates(gr, chain_file)
}
\arguments{
\item{gr}{a `GRanges` object}

\item{chain_file}{A path to the chain file}
}
\value{
A `GRanges` object with coordinates lifted over to the new assembly
}
\description{
Performs liftOver to convert genomic coordinates from one reference genome 
assembly to another (e.g., from hg19 to hg38). This is essential when working 
with data from different genome builds or when updating analyses to newer 
genome versions. Uses UCSC chain files for accurate coordinate mapping. Some 
regions may not lift over if they are in areas that differ significantly 
between genome builds.
}
\examples{
bed_file <- system.file("extdata", "example.bed", package = "ReAnnotateR")
chain_file <- system.file("extdata", "hg19ToHg38.over.chain", package = "ReAnnotateR")
if (file.exists(bed_file) && file.exists(chain_file)) {
  gr <- read_bed(bed_file)
  gr_new <- convert_coordinates(gr, chain_file)
}
}
