% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{functional_terms}
\alias{functional_terms}
\title{Retrieve functional annotations for genes}
\usage{
functional_terms(gr, orgdb, go = TRUE, kegg = FALSE)
}
\arguments{
\item{gr}{GRanges object with a metadata column 'nearest_gene' of gene IDs}

\item{orgdb}{An OrgDb object for mapping gene IDs to GO terms}

\item{go}{TRUE/FALSE, if TRUE, perform GO enrichment (default TRUE)}

\item{kegg}{TRUE/FALSE, if TRUE, perform KEGG pathway enrichment (default FALSE)}
}
\value{
A list with elements 'go' and/or 'kegg', each containing an enrichment result data frame
}
\description{
For each gene associated with genomic intervals, retrieve functional
information
}
\examples{
library(GenomicRanges)
library(clusterProfiler)
library(org.Hs.eg.db)
gr <- GRanges(seqnames = "chr1", ranges = IRanges::IRanges(100000, 101000))
GenomicRanges::mcols(gr)$nearest_gene <- "1"
# Note: Running this example requires a valid OrgDb object like org.Hs.eg.db
# functional_terms(gr, org.Hs.eg.db)

}
