% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot Functional Enrichment Results}
\usage{
plot_enrichment(enrich_result, top_n = 10)
}
\arguments{
\item{enrich_result}{An object returned from \code{functional_terms()}}

\item{top_n}{Integer which is the number of top terms to plot (default 10)}
}
\value{
A ggplot2 object showing enrichment scores for top terms
}
\description{
Visualizes enriched biological functions or pathways.
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
enrich_res <- functional_terms(gr)
plot_enrichment(enrich_res)
}
