% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_feature_composition}
\alias{plot_feature_composition}
\title{Plot Feature Composition of Genomic Intervals}
\usage{
plot_feature_composition(gr)
}
\arguments{
\item{gr}{A GRanges object with a metadata column named "feature" showing the feature type for each interval}
}
\value{
A ggplot2 object showing feature composition as a bar plot.
}
\description{
Plots the proportion of genomic intervals that overlap promoters, exons, introns, or intergenic regions
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
gr <- annotate_regions(gr, txdb_info)
plot_feature_composition(gr)
}
