% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_missing_annotations}
\alias{plot_missing_annotations}
\title{Plot Missing Annotations}
\usage{
plot_missing_annotations(gr)
}
\arguments{
\item{gr}{GRanges object with a "feature_type" metadata column}
}
\value{
ggplot2 barplot showing counts of annotated vs. unannotated intervals
}
\description{
Creates a bar plot showing annotated vs missing intervals
}
\examples{
library(GenomicRanges)
gr <- GRanges(seqnames = "chr1", ranges = IRanges::IRanges(c(100, 200, 300, 400), 
                                                            c(150, 250, 350, 450)))
mcols(gr)$feature_type <- c("promoter", "intergenic", NA, "exon")
plot_missing_annotations(gr)
}
