% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_regulatory_regions}
\alias{plot_regulatory_regions}
\title{Plot Regulatory Regions Overlaps}
\usage{
plot_regulatory_regions(gr, regulatory_gr)
}
\arguments{
\item{gr}{GRanges object with your intervals}

\item{regulatory_gr}{GRanges object with regulatory regions}
}
\value{
ggplot2 object showing which intervals overlap regulatory regions
}
\description{
Creates a bar plot showing overlaps with regulatory regions
}
\examples{
\dontrun{
bed_file <- system.file("extdata", "example.bed", package = "ReAnnotateR")
gr <- read_bed(bed_file)
# regulatory_gr would be loaded from a data source
# plot_regulatory_regions(gr, regulatory_gr)
}
}
