% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReAnnotateR-package.R
\docType{package}
\name{ReAnnotateR-package}
\alias{ReAnnotateR}
\alias{ReAnnotateR-package}
\title{ReAnnotateR: Re-annotate Genomic Intervals from BED Files}
\description{
Provides functions to annotate genomic regions from BED files by linking intervals to genes, genomic features, and biological functions. Includes visualization and enrichment analysis.
}
\author{
\strong{Maintainer}: Julia Gilley \email{julia.gilley@mail.utoronto.ca}

}
\keyword{internal}
