% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runReAnnotateR.R
\name{runReAnnotateR}
\alias{runReAnnotateR}
\title{Launch ReAnnotateR Shiny App}
\usage{
runReAnnotateR()
}
\value{
Launches the Shiny application
}
\description{
Launches the interactive Shiny application for genomic region annotation
}
\examples{
\dontrun{
runReAnnotateR()
}
}
