% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{annotate_regions}
\alias{annotate_regions}
\title{Annotate each genomic region with its feature type}
\usage{
annotate_regions(gr, txdb_info, promoter_up = 3000, promoter_down = 3000)
}
\arguments{
\item{gr}{GRanges object of genomic regions}

\item{txdb_info}{A TxDb object with gene coordinates on a certain reference genome}

\item{promoter_up}{Optionally, the number of bases upstream of a transcription start site to include in the promoter region (default 3000)}

\item{promoter_down}{Optionally, the number of bases downstream of a transcription start site to include in the promoter region (default 3000)}
}
\value{
A GRanges object that includes a new column called 'feature_type' showing which type of genomic feature each region overlaps
}
\description{
Assigns the genomic feature of promoter, exon, intron, or intergenic to each interval
}
\examples{
library(GenomicRanges)
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
gr <- GRanges(seqnames = "chr1", ranges = IRanges(100000, 101000))
annotate_regions(gr, txdb)
}
