% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{convert_coordinates}
\alias{convert_coordinates}
\title{Convert genomic coordinates between assemblies}
\usage{
convert_coordinates(gr, chain_file)
}
\arguments{
\item{gr}{a `GRanges` object}

\item{chain_file}{A path to the chain file}
}
\value{
A `GRanges` object with coordinates lifted over to the new assembly
}
\description{
Convert genomic coordinates between assemblies
}
\examples{
bed_file <- system.file("extdata", "example.bed", package = "ReAnnotateR")
gr <- read_bed(bed_file)
chain_file <- system.file("extdata", "hg19ToHg38.over.chain", package = "ReAnnotateR")
gr_new <- convert_coordinates(gr, chain_file)
}
