% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{convert_coordinates}
\alias{convert_coordinates}
\title{Convert genomic coordinates between assemblies}
\usage{
convert_coordinates(gr, chain_file)
}
\arguments{
\item{gr}{a `GRanges` object}

\item{chain_file}{A path to the chain file}
}
\value{
A `GRanges` object with coordinates lifted over to the new assembly
}
\description{
Uses a chain file to lift genomic coordinates from one reference assembly to another
}
\examples{
bed_file <- system.file("extdata", "example.bed", package = "ReAnnotateR")
chain_file <- system.file("extdata", "hg19ToHg38.over.chain", package = "ReAnnotateR")
if (file.exists(bed_file) && file.exists(chain_file)) {
  gr <- read_bed(bed_file)
  gr_new <- convert_coordinates(gr, chain_file)
}
}
