% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{export_results}
\alias{export_results}
\title{Export annotated or enriched results}
\usage{
export_results(object, file, format = "tsv")
}
\arguments{
\item{object}{The data to export (annotated GRanges or enrichment data frame)}

\item{file}{Path to the output file}

\item{format}{Output format, either "tsv" or "csv"}
}
\value{
None because writes a file to disk
}
\description{
Saves annotated genomic intervals or enrichment results to a file
Supports writing GRanges, data frames, or tibbles in TSV or CSV format
}
\examples{
bed_path <- system.file("extdata", "example.bed", package = "ReAnnotateR")
gr <- read_bed(bed_path)
export_results(gr, tempfile(fileext = ".tsv"), format = "tsv")
}
