% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{fisher_enrichment}
\alias{fisher_enrichment}
\title{Perform Fisher's exact test for enrichment!}
\usage{
fisher_enrichment(gr, category, background_gr)
}
\arguments{
\item{gr}{GRanges object with a metadata column of categories (like feature)}

\item{category}{A character string specifying which category to test for enrichment. Valid options can be any value present in the metadata column. So, if using the 'feature' column, valid options are: "promoter", "exon", "intron", "intergenic".}

\item{background_gr}{A GRanges object representing the background set of regions.}
}
\value{
A list with elements 'odds_ratio', 'p_value', and 'table' containing the contingency table, or NULL if the category is invalid
}
\description{
Performs a Fisher's exact test to see if a category (like a promoter) is overrepresented among a set of genomic regions compared to a background set or not
}
\examples{
library(GenomicRanges)
gr <- GRanges(seqnames = "chr1", ranges = IRanges(c(100,200), c(150,250)))
mcols(gr)$feature <- c("promoter", "exon")
bg_gr <- GRanges(seqnames = "chr1", ranges = IRanges(c(50,150,250), c(120,220,300)))
mcols(bg_gr)$feature <- c("promoter", "exon", "intron")
fisher_enrichment(gr, "promoter", bg_gr)
}
