% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{nearest_gene}
\alias{nearest_gene}
\title{Identify Nearest Genes and Calculate Distances}
\usage{
nearest_gene(gr, txdb_info)
}
\arguments{
\item{gr}{GRanges object of genomic regions}

\item{txdb_info}{A TxDb object with gene coordinates on a certain reference genome}
}
\value{
A GRanges object with extra metadata columns 'nearest_gene' and 'distance_to_gene'
}
\description{
For each genomic interval, identifies the closest gene (by distance from
interval to gene body) and calculates the genomic distance in base pairs.
A distance of 0 indicates the interval overlaps the gene body. Positive
distances indicate how far the interval is from the nearest gene. This
information is crucial for linking regulatory elements to their potential
target genes or for understanding the genomic context of your intervals.
Must be run after annotate_regions().
}
\examples{
\dontrun{
# Requires TxDb.Hsapiens.UCSC.hg19.knownGene (not available on all systems)
library(GenomicRanges)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
gr <- GRanges(seqnames = "chr1", ranges = IRanges(100000, 101000))
gr_annotated <- annotate_regions(gr, txdb)
nearest_gene(gr_annotated, txdb)
}
}
