% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{nearest_gene}
\alias{nearest_gene}
\title{Find the closest gene to each genomic region}
\usage{
nearest_gene(gr, txdb_info)
}
\arguments{
\item{gr}{GRanges object of genomic regions}

\item{txdb_info}{A TxDb object with gene coordinates on a certain reference genome}
}
\value{
A GRanges object with extra metadata columns 'nearest_gene' and 'distance_to_gene'
}
\description{
For each genomic interval, this function identifies the nearest gene and calculates the distance
from the interval to the gene
}
\examples{
library(GenomicRanges)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txdb <- TxDb.Hsapiens.UCSC.hg19.knownGene
gr <- GRanges(seqnames = "chr1", ranges = IRanges(100000, 101000))
nearest_gene(gr, txdb)
}
