% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_annotation_summary}
\alias{plot_annotation_summary}
\title{Plot Annotation Summary}
\usage{
plot_annotation_summary(gr)
}
\arguments{
\item{gr}{GRanges object with a "feature_type" metadata column}
}
\value{
ggplot2 barplot showing counts of intervals per feature type
}
\description{
Creates a bar plot summarizing all feature types
}
\examples{
library(GenomicRanges)
gr <- GRanges(seqnames = "chr1", ranges = IRanges::IRanges(c(100, 200, 300, 400, 500), 
                                                            c(150, 250, 350, 450, 550)))
mcols(gr)$feature_type <- c("promoter", "exon", "intron", "intergenic", NA)
plot_annotation_summary(gr)
}
