% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_annotation_summary}
\alias{plot_annotation_summary}
\title{Plot Annotation Summary}
\usage{
plot_annotation_summary(gr)
}
\arguments{
\item{gr}{GRanges object with a "feature" metadata column}
}
\value{
ggplot2 barplot showing counts of intervals per feature type
}
\description{
Plot Annotation Summary
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
gr$feature <- c("promoter", "exon", "intron", "intergenic", NA)
plot_annotation_summary(gr)
}
