% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_chromosomal_density}
\alias{plot_chromosomal_density}
\title{Plot Chromosomal Density of Genomic Intervals}
\usage{
plot_chromosomal_density(gr, bin_size = 1e+06)
}
\arguments{
\item{gr}{A GRanges object containing genomic intervals}

\item{bin_size}{Integer specifying the size of bins in base pairs (default 1e6)}
}
\value{
A ggplot2 object showing the number of intervals per bin along each chromosome
}
\description{
Shows the distribution of genomic intervals along chromosomes
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
plot_chromosomal_density(gr)
}
