% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_chromosomal_density}
\alias{plot_chromosomal_density}
\title{Plot Chromosomal Density of Genomic Intervals}
\usage{
plot_chromosomal_density(gr, bin_size = 1e+06)
}
\arguments{
\item{gr}{GRanges object containing genomic intervals}

\item{bin_size}{Integer specifying bin size in base pairs (default 1e6)}
}
\value{
ggplot2 object showing number of intervals per bin along each chromosome
}
\description{
Creates a line plot showing the density of intervals across chromosomes
}
\examples{
bed_file <- system.file("extdata", "example.bed", package = "ReAnnotateR")
if (file.exists(bed_file)) {
  gr <- read_bed(bed_file)
  plot_chromosomal_density(gr)
}
}
