% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot Functional Enrichment Results}
\usage{
plot_enrichment(enrich_result, top_n = 10)
}
\arguments{
\item{enrich_result}{Object returned from functional_terms()}

\item{top_n}{Integer, number of top terms to plot (default 10)}
}
\value{
ggplot2 object showing enrichment scores for top terms
}
\description{
Plot Functional Enrichment Results
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
enrich_res <- functional_terms(gr)
plot_enrichment(enrich_res)
}
