% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plot Functional Enrichment Results}
\usage{
plot_enrichment(enrich_result, top_n = 10)
}
\arguments{
\item{enrich_result}{Data frame or enrichment result object returned from functional_terms()}

\item{top_n}{Integer, number of top terms to plot (default 10)}
}
\value{
ggplot2 object showing enrichment scores for top terms
}
\description{
Creates a bar plot of the top enriched terms
}
\examples{
\dontrun{
bed_file <- system.file("extdata", "example.bed", package = "ReAnnotateR")
gr <- read_bed(bed_file)
enrich_res <- functional_terms(gr, org.Hs.eg.db)
plot_enrichment(enrich_res$go)
}
}
