% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_feature_composition}
\alias{plot_feature_composition}
\title{Plot Feature Composition of Genomic Intervals}
\usage{
plot_feature_composition(gr)
}
\arguments{
\item{gr}{GRanges object with a "feature_type" metadata column}
}
\value{
ggplot2 object showing feature composition as a bar plot
}
\description{
Creates a bar plot showing the distribution of genomic features
}
\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
bed_file <- system.file("extdata", "example.bed", package = "ReAnnotateR")
gr <- read_bed(bed_file)
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
gr_annotated <- annotate_regions(gr, txdb)
plot_feature_composition(gr_annotated)
}
}
