% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_feature_composition}
\alias{plot_feature_composition}
\title{Visualize Genomic Feature Type Distribution}
\usage{
plot_feature_composition(gr)
}
\arguments{
\item{gr}{GRanges object with a "feature_type" metadata column}
}
\value{
ggplot2 object showing feature composition as a bar plot
}
\description{
Creates a bar plot showing the count of genomic intervals classified into
each feature type: promoters, exons, introns, and intergenic regions. This
provides a quick overview of which types of genomic elements your intervals
predominantly overlap. For example, if most intervals fall in promoters, your
dataset may be enriched for regulatory regions near transcription start sites.
}
\examples{
\dontrun{
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
bed_file <- system.file("extdata", "example.bed", package = "ReAnnotateR")
gr <- read_bed(bed_file)
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
gr_annotated <- annotate_regions(gr, txdb)
plot_feature_composition(gr_annotated)
}
}
