% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_ideogram}
\alias{plot_ideogram}
\title{Plot Chromosome Ideogram with Intervals}
\usage{
plot_ideogram(gr, genome = "hg38", chromosomes = NULL)
}
\arguments{
\item{gr}{GRanges object with genomic intervals}

\item{genome}{Reference genome build (default "hg38")}

\item{chromosomes}{Character vector of chromosomes to plot (default all)}
}
\value{
Gviz plot displaying intervals on chromosome ideograms
}
\description{
Plot Chromosome Ideogram with Intervals
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
plot_ideogram(gr)
}
