% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_ideogram}
\alias{plot_ideogram}
\title{Plot Chromosome Ideogram with Intervals}
\usage{
plot_ideogram(gr, genome = "hg38", chromosomes = NULL)
}
\arguments{
\item{gr}{GRanges object with genomic intervals}

\item{genome}{Reference genome build (default "hg38")}

\item{chromosomes}{Character vector of chromosomes to plot (default all)}
}
\value{
NULL (plots are created as side effect)
}
\description{
Creates an ideogram visualization showing intervals on chromosomes
}
\examples{
\dontrun{
bed_file <- system.file("extdata", "example.bed", package = "ReAnnotateR")
gr <- read_bed(bed_file)
plot_ideogram(gr)
}
}
