% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_missing_annotations}
\alias{plot_missing_annotations}
\title{Plot Missing Annotations}
\usage{
plot_missing_annotations(gr)
}
\arguments{
\item{gr}{GRanges object with a "feature" metadata column}
}
\value{
ggplot2 barplot showing counts of annotated vs. unannotated intervals
}
\description{
Plot Missing Annotations
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
gr$feature <- c("promoter", "intergenic", NA, "exon")
plot_missing_annotations(gr)
}
