% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_missing_annotations}
\alias{plot_missing_annotations}
\title{Plot Missing Annotations}
\usage{
plot_missing_annotations(gr)
}
\arguments{
\item{gr}{A GRanges object with a metadata column \code{feature} indicating annotation.}
}
\value{
A ggplot2 barplot showing counts of annotated vs. unannotated intervals.
}
\description{
Shows intervals that could not be annotated or mapped to features/genes.
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
gr$feature <- c("promoter", "intergenic", NA, "exon")
plot_missing_annotations(gr)
}
