% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_regulatory_regions}
\alias{plot_regulatory_regions}
\title{Plot Regulatory Regions Overlaps}
\usage{
plot_regulatory_regions(gr, regulatory_gr)
}
\arguments{
\item{gr}{A GRanges object with your intervals.}

\item{regulatory_gr}{A GRanges object with regulatory regions.}
}
\value{
A ggplot2 object showing which intervals overlap regulatory regions.
}
\description{
Highlights intervals overlapping regulatory elements like enhancers or promoters.
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
# regulatory_gr would be your regulatory region GRanges
plot_regulatory_regions(gr, regulatory_gr)
}
