% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_regulatory_regions}
\alias{plot_regulatory_regions}
\title{Plot Regulatory Regions Overlaps}
\usage{
plot_regulatory_regions(gr, regulatory_gr)
}
\arguments{
\item{gr}{GRanges object with your intervals}

\item{regulatory_gr}{GRanges object with regulatory regions}
}
\value{
ggplot2 object showing which intervals overlap regulatory regions
}
\description{
Plot Regulatory Regions Overlaps
}
\examples{
gr <- read_bed(system.file("extdata", "example.bed", package = "ReAnnotateR"))
plot_regulatory_regions(gr, regulatory_gr)
}
