% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_bed}
\alias{read_bed}
\title{Get a GRanges object from a BED file}
\usage{
read_bed(file)
}
\arguments{
\item{file}{path to a BED file}
}
\value{
\code{GRanges} object containing the intervals from the BED file
}
\description{
Checks for invalid chromosome start and end positions
}
\examples{
data(example_bed)
bed_path <- system.file("extdata", "example.bed", package = "ReAnnotateR")
gr <- read_bed(bed_path)
}
