% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_bed}
\alias{read_bed}
\title{Read BED file into GRanges object}
\usage{
read_bed(file)
}
\arguments{
\item{file}{path to a BED file}
}
\value{
`GRanges` object containing the intervals from the BED file
}
\description{
Loads a BED file and converts it to a GRanges object.
Checks for invalid chromosome start and end positions.
}
\examples{
bed_path <- system.file("extdata", "example.bed", package = "ReAnnotateR")
if (file.exists(bed_path)) {
  gr <- read_bed(bed_path)
}
}
